refactor: Move ParseResult to bpdm-common for shared use..
