+++ bpdm-pool/src/main/kotlin/org/eclipse/tractusx/bpdm/pool/entity/ConfigurationEntryDb.kt
