+++ bpdm-pool/src/main/kotlin/org/eclipse/tractusx/bpdm/pool/exception/BpdmRequestSizeException.kt
