+++ bpdm-pool/src/main/kotlin/org/eclipse/tractusx/bpdm/pool/service/application/v6/AddressSearchApplicationV6Service.kt
@@ -28,7 +28,7 @@ import org.eclipse.tractusx.bpdm.pool.api.v6.model.LogisticAddressVerboseDtoV6
 import org.eclipse.tractusx.bpdm.pool.api.v6.model.request.AddressSearchRequestV6
 import org.eclipse.tractusx.bpdm.pool.mapper.poolv6.inbound.AddressSearchRequestMapperV6
 import org.eclipse.tractusx.bpdm.pool.mapper.poolv6.outbound.AddressResponseMapperV6
-import org.eclipse.tractusx.bpdm.pool.model.ParseResult
+import org.eclipse.tractusx.bpdm.common.model.ParseResult
 import org.eclipse.tractusx.bpdm.pool.service.operation.address.AddressSearchService
 import org.eclipse.tractusx.bpdm.pool.service.parser.address.AddressSearchParser
 import org.eclipse.tractusx.bpdm.pool.service.parser.legalentity.LegalEntityAddressSearchParser
