+++ bpdm-pool/src/main/kotlin/org/eclipse/tractusx/bpdm/pool/service/application/v6/ChangelogSearchApplicationV6Service.kt
@@ -28,7 +28,7 @@ import org.eclipse.tractusx.bpdm.pool.api.v6.model.response.ChangelogEntryVerbos
 import org.eclipse.tractusx.bpdm.pool.mapper.poolv6.inbound.ChangelogSearchRequestMapperV6
 import org.eclipse.tractusx.bpdm.pool.mapper.poolv6.outbound.ChangelogResponseMapperV6
 import org.eclipse.tractusx.bpdm.pool.mapper.shared.outbound.ChangelogParseErrorMapper
-import org.eclipse.tractusx.bpdm.pool.model.ParseResult
+import org.eclipse.tractusx.bpdm.common.model.ParseResult
 import org.eclipse.tractusx.bpdm.pool.service.operation.changelog.ChangelogSearchService
 import org.eclipse.tractusx.bpdm.pool.service.parser.ChangelogSearchParser
 import org.springframework.stereotype.Service
