+++ bpdm-pool/src/main/kotlin/org/eclipse/tractusx/bpdm/pool/service/parser/ChangelogSearchParser.kt
@@ -20,7 +20,7 @@
 package org.eclipse.tractusx.bpdm.pool.service.parser
 
 import org.eclipse.tractusx.bpdm.pool.config.ControllerConfigProperties
-import org.eclipse.tractusx.bpdm.pool.model.ParseResult
+import org.eclipse.tractusx.bpdm.common.model.ParseResult
 import org.eclipse.tractusx.bpdm.pool.model.error.ChangelogSearchParseError
 import org.eclipse.tractusx.bpdm.pool.model.error.SearchValuesTooMany
 import org.eclipse.tractusx.bpdm.pool.model.parsed.ChangelogSearchParsed
