+++ bpdm-pool/src/main/kotlin/org/eclipse/tractusx/bpdm/pool/service/parser/bpn/BpnIdentifierSearchParser.kt
@@ -20,7 +20,7 @@
 package org.eclipse.tractusx.bpdm.pool.service.parser.bpn
 
 import org.eclipse.tractusx.bpdm.pool.config.ControllerConfigProperties
-import org.eclipse.tractusx.bpdm.pool.model.ParseResult
+import org.eclipse.tractusx.bpdm.common.model.ParseResult
 import org.eclipse.tractusx.bpdm.pool.model.error.BpnIdentifierSearchParseError
 import org.eclipse.tractusx.bpdm.pool.model.error.SearchValuesTooMany
 import org.eclipse.tractusx.bpdm.pool.model.parsed.BpnIdentifierSearchParsed
