+++ bpdm-pool/src/main/kotlin/org/eclipse/tractusx/bpdm/pool/service/parser/legalentity/LegalEntitySiteSearchParser.kt
@@ -19,7 +19,7 @@
 
 package org.eclipse.tractusx.bpdm.pool.service.parser.legalentity
 
-import org.eclipse.tractusx.bpdm.pool.model.ParseResult
+import org.eclipse.tractusx.bpdm.common.model.ParseResult
 import org.eclipse.tractusx.bpdm.pool.model.error.UnresolvableLegalEntity
 import org.eclipse.tractusx.bpdm.pool.model.parsed.SiteSearchParsed
 import org.eclipse.tractusx.bpdm.pool.model.request.SiteSearchRequest
